requirements.txt
allauth-django==0.42.6
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.6
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.7.1
Deprecated==1.2.12
distlib==0.3.1
Django==3.2.4
django-allauth==0.44.0
django-braces==1.14.0
django-cors-headers==3.7.0
django-countries==7.1
django-dotenv==1.4.2
django-oauth-toolkit==1.5.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
facebook-sdk==3.1.0
filelock==3.0.12
html5lib==1.1
idna==2.10
jwcrypto==0.9.1
lxml==4.6.3
mysqlclient==2.0.3
numpy==1.21.0
oauthlib==3.1.0
pandas==1.3.0
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
virtualenv==20.4.3
webencodings==0.5.1
wrapt==1.12.1
